Refactor browser sidecar modules
